fixed bug in intervalSet in which the time points could be incorrectly saved as int32s instead of doubles, which crashes matlab when the mex file is called

+++ @intervalSet/intervalSet.m
@@ -169,8 +169,8 @@ function O = intervalSet(varargin)
       cnvrt = 1;
   end
   
-  O.start = cnvrt * s;
-  O.stop = cnvrt * e;
+  O.start = double(cnvrt * s);
+  O.stop = double(cnvrt * e);
   O.units = units('time', 'ts');
   O = class(O, 'intervalSet');
